% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bubbleplot_pwys.R
\name{bubbleplot_pwys}
\alias{bubbleplot_pwys}
\title{Make bubble plots for the significance, DE proportion, and counts of pathway analysis}
\usage{
bubbleplot_pwys(
  tab,
  prefix.v = NULL,
  name = NA,
  type.sig = c("p", "FDR"),
  cut.sig = 0.05,
  ntop = 20,
  pwys_nm_size = 100,
  width = 8,
  height = 8,
  caption = TRUE,
  mixed = c("include", "exclude", "only")
)
}
\arguments{
\item{tab}{Table of output from \code{ezlimma::roast_*}.}

\item{prefix.v}{A vector of prefixes that prefix \code{.p}, \code{.FDR}, or \code{.logFC} in \code{colnames(tab)}.
These will be the comparisons shown in the plot. If \code{NULL} these are inferred from \code{colnames(tab)} that end with \code{.p}.}

\item{name}{Name of file to create. Set to \code{NA} to plot to screen instead of to file; otherwise, "_bubbleplots.pdf" is appended to the name.}

\item{type.sig}{Either "p" or "FDR"; type of significance to show.}

\item{cut.sig}{Numeric in [0, 1]. Pathways need to have significance of type \code{type.sig < cut.sig} in a comparison to be shown on the dot plot.}

\item{ntop}{Integer; number of top pathways to show.}

\item{pwys_nm_size}{The maximum number of characters for pathway names. Longer names will be truncated.}

\item{width}{Manual option for determining the output file width in inches.}

\item{height}{Manual option for determining the output file height in inches.}

\item{caption}{Logical; should the caption explaining the x-axis be included?}

\item{mixed}{Character string. Should mixed statistics be included, should they be the only statistics included, or should they be excluded?}
}
\value{
Invisibly, a list of ggplot objects.
}
\description{
Make bubble plots like clusterProfiler showing the significance, proportion of genes with p-value below 0.05, and counts per comparison
per pathway. These are called "bubble plots" instead of "dot plots"  because they represent the counts as an additional dimension of
numerical (i.e. non-factor) data.
}
