% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dotplot_pwys.R
\name{dotplot_pwys}
\alias{dotplot_pwys}
\title{Make dot plots for the significance and DE proportion of pathway analysis}
\usage{
dotplot_pwys(
  tab,
  prefix.v = NULL,
  name = NA,
  type.sig = c("p", "FDR"),
  cut.sig = 0.05,
  ntop = 20,
  pwys_nm_size = 100,
  font.size = 12,
  width = 8,
  height = 8,
  mixed = c("include", "exclude", "only"),
  reorder.rows = TRUE,
  caption = TRUE,
  colorbar.title = "Prop P<5\%"
)
}
\arguments{
\item{tab}{Table of output from \code{ezlimma::roast_*}.}

\item{prefix.v}{A vector of prefixes that prefix \code{.p}, \code{.FDR}, or \code{.logFC} in \code{colnames(tab)}.
These will be the comparisons shown in the plot. If \code{NULL} these are inferred from \code{colnames(tab)} that end with \code{.p}.}

\item{name}{Name of file to create. Set to \code{NA} to plot to screen instead of to file; otherwise, "_dotplot.pdf" is appended to the name.}

\item{type.sig}{Either "p" or "FDR"; type of significance to show.}

\item{cut.sig}{Numeric in [0, 1]. Pathways need to have significance of type \code{type.sig < cut.sig} in a comparison to be shown on the dot plot.}

\item{ntop}{Integer; number of top pathways to show.}

\item{pwys_nm_size}{The maximum number of characters for pathway names. Longer names will be truncated.}

\item{font.size}{Numeric font size of pathway names e.g. 12.}

\item{width}{Manual option for determining the output file width in inches.}

\item{height}{Manual option for determining the output file height in inches.}

\item{mixed}{Character string. Should mixed statistics be included, should they be the only statistics included, or should they be excluded?}

\item{reorder.rows}{Boolean; reorder pathways by significance before selecting \code{ntop} pathways? If not, order of \code{tab} is retained.}

\item{caption}{Logical; should the caption explaining the color bar title be included?}

\item{colorbar.title}{Character title of color bar.}
}
\value{
Invisibly, the ggplot object.
}
\description{
Make dot plots like clusterProfiler showing the significance and proportion of genes in each pathway
with significance below cutoff.
}
