% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_prefix.R
\name{extract_prefix}
\alias{extract_prefix}
\title{Extract prefix from names}
\usage{
extract_prefix(nm, suffix, sep = ".")
}
\arguments{
\item{nm}{Character vector of names to search.}

\item{suffix}{Character string to search for at end of \code{cnames}.}

\item{sep}{Separator in \code{cnames} preceeding \code{suffix}. If \code{NA}, elements of \code{nm} must
match \code{suffix} exactly.}
}
\value{
A character vector of prefixes. If there is no prefix and no sep, \code{NA} is returned.
}
\description{
Extract prefixes from names, presumably colnames.
}
