% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ezpca.R
\name{ezpca}
\alias{ezpca}
\title{PCA plot of top two components}
\usage{
ezpca(object, pheno, name = "pca", scale. = FALSE, alpha = 1,
  all.size = NULL, facet = NULL, rm.leg.title = FALSE, labels = FALSE,
  manual.color = NULL, manual.shape = NULL, ...)
}
\arguments{
\item{object}{Matrix-like object with samples as columns.}

\item{pheno}{Dataframe with rows as samples and columns as phenotypes.}

\item{name}{Name of PNG to write out. Set to \code{NA} to suppress writing to file.}

\item{scale.}{Logical indicating whether to scale in \code{\link[stats]{prcomp}}.}

\item{alpha}{Transparency, passed to \code{\link[ggplot2]{geom_point}}.}

\item{all.size}{Passed to \code{\link[ggplot2]{geom_point}} \code{size} parameter to give size for all points without
appearing in legend. \code{ggplot2} default is size=2.}

\item{facet}{A formula with columns in \code{pheno} to facet by.}

\item{rm.leg.title}{Logical indicating if legend title should be removed.}

\item{labels}{Logical indicating if sample labels should be added next to points.}

\item{manual.color}{Vector passed to \code{\link[ggplot2]{scale_colour_manual}}.}

\item{manual.shape}{Vector passed to \code{\link[ggplot2]{scale_shape_manual}}.}

\item{...}{Passed to \code{\link[ggplot2]{ggplot}} \code{aes_string} parameter.}
}
\description{
PCA plot of top two principal components using \code{ggplot2}.
}
