% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eztsne.R
\name{eztsne}
\alias{eztsne}
\title{tSNE plot of first two dimensions}
\usage{
eztsne(object, pheno.df, name = "tsne", check_duplicates = FALSE,
  pca = TRUE, perplexity = min(30, round(ncol(object)/5)),
  theta = 0.5, seed = 123, initial_dims = 50, max_iter = 1000,
  alpha = 1, all.size = NULL, facet = NULL, title = NULL,
  subtitle = NULL, rm.leg.title = FALSE, labels = FALSE,
  manual.color = NULL, manual.shape = NULL, plot = TRUE, ...)
}
\arguments{
\item{object}{Matrix-like object with features (e.g. genes) as rows and samples as columns.}

\item{pheno.df}{Data frame with rows as samples and columns as phenotypes.}

\item{name}{Name of file to create. Set to \code{NA} to plot to screen instead of to file.}

\item{check_duplicates}{logical; Checks whether duplicates are present. It is best to make sure there are no duplicates present and set this option to FALSE, especially for large datasets (default: TRUE)}

\item{pca}{logical; Whether an initial PCA step should be performed (default: TRUE)}

\item{perplexity}{numeric; Perplexity parameter (should not be bigger than 3 * perplexity < nrow(X) - 1, see details for interpretation)}

\item{theta}{numeric; Speed/accuracy trade-off (increase for less accuracy), set to 0.0 for exact TSNE (default: 0.5)}

\item{seed}{Integer seed to set for reproducility if \code{fun="mroast"}, since \code{mroast} uses random 
simulations. Ignored if \code{fun="fry"}.}

\item{initial_dims}{integer; the number of dimensions that should be retained in the initial PCA step (default: 50)}

\item{max_iter}{integer; Number of iterations (default: 1000)}

\item{alpha}{Transparency, passed to \code{\link[ggplot2]{geom_point}}.}

\item{all.size}{Passed to \code{\link[ggplot2]{geom_point}} \code{size} parameter to give size for all points without
appearing in legend. \code{ggplot2} default is size=2.}

\item{facet}{A formula with columns in \code{pheno.df} to facet by.}

\item{title}{Title text; suppressed if it is \code{NULL}.}

\item{subtitle}{Subtitle text; suppressed if it is \code{NULL} or \code{title} is \code{NULL}. If you'd like a
\code{subtitle} but no \code{title}, set \code{title = ""}.}

\item{rm.leg.title}{Logical indicating if legend title should be removed.}

\item{labels}{Logical, should sample labels be added next to points?}

\item{manual.color}{Vector passed to \code{\link[ggplot2:scale_manual]{scale_colour_manual}}.}

\item{manual.shape}{Vector passed to \code{\link[ggplot2:scale_manual]{scale_shape_manual}}.}

\item{plot}{Logical; should plot be generated?}

\item{...}{Passed to \code{\link[ggplot2:aes_]{aes_string}}.}
}
\value{
Invisibly, a \code{ggplot} object. Its \code{data} element contains the first two principal components
appended to \code{pheno.df}.
}
\description{
tSNE plot of first two dimensions using \pkg{ggplot2}.
}
\details{
\code{object} must have colnames, and if \code{pheno.df}
is given, it is checked that \code{colnames(object)==rownames(pheno.df)}.
}
