% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ezvenn.R
\name{ezvenn}
\alias{ezvenn}
\title{Venn diagrams}
\usage{
ezvenn(tab, prefix.v, fdr.cutoff = NULL, p.cutoff = NULL,
  logfc.cutoff = NULL, circle.names = prefix.v, main = "", name = NA,
  cex = c(1, 1, 1), plot = TRUE)
}
\arguments{
\item{tab}{Table of output from \code{ezlimma}.}

\item{prefix.v}{A vector of prefixes that prefix \code{.p}, \code{.FDR}, or \code{.logFC} in \code{colnames(tab)}.}

\item{fdr.cutoff}{FDR cutoff to threshold features in \code{tab}. FDRs < \code{fdr.cutoff} are significant.}

\item{p.cutoff}{p-value cutoff to threshold features in \code{tab}. p-values < \code{p.cutoff} are significant.}

\item{logfc.cutoff}{log fold-change cutoff to threshold features in \code{tab}, when \code{tab} came fro \code{\link[ezlimma]{limma_contrasts}}.}

\item{circle.names}{Names of circles corresponding to \code{prefix.v}.}

\item{main}{Main title of plot.}

\item{name}{Name of PDF that gets "_venn" appended, then written. Set to \code{NA} to suppress writing to file.}

\item{cex}{A numerical value giving the amount by which plotting text and symbols should be magnified relative to the default. See \code{\link[graphics]{par}}.}

\item{plot}{If the Venn diagram should be plotted.}
}
\value{
Binary matrix indicating which features (rows) of \code{tab} were significant with specified cutoffs.

A ggplot object, invisibly.
}
\description{
Venn diagrams using output from \code{ezlimma} package.
}
\details{
One of \code{fdr.cutoff} or \code{p.cutoff} must be given. If both are given, only \code{fdr.cutoff} is used. \code{logfc.cutoff} if given is used in addition to these.
}
