% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ezvolcano.R
\name{ezvolcano}
\alias{ezvolcano}
\title{Volcano plot in ggplot2}
\usage{
ezvolcano(tab, lfc.col = NULL, sig.col = NULL, lab.col = NULL,
  ntop.sig = 0, ntop.lfc = 0, comparison = NULL, name = "volcano",
  ann.rnames = NULL, up.color = "black", down.color = "black",
  x.bound = NULL, y.bound = NULL, type.sig = c("p", "FDR"),
  cut.color = NULL, cut.lfc = 1, cut.sig = 0.05, sep = ".",
  na.lab = c("---", ""))
}
\arguments{
\item{tab}{Table of output from \code{ezlimma}.}

\item{lfc.col}{Column name or index of tab with logFC. Some features should be > 0 and others < 0.}

\item{sig.col}{Column name or index of tab with p-values or FDRs.}

\item{lab.col}{Column name or index of tab with labels, such as gene symbol, annotating features. If \code{NULL},
no points are annotated.}

\item{ntop.sig}{Number of top significant features to annotate.}

\item{ntop.lfc}{Number of top logFC features to annotate.}

\item{comparison}{Name of contrast to plot. If given, it's assumed that \code{lfc.col=paste0(comparison, '.logFC')}
and \code{sig.col=paste0(comparison, '.p') or paste0(comparison, '.FDR')}, and these are over-ridden.}

\item{name}{Name of PNG file to write to. Set to \code{NA} to suppress writing to file.}

\item{ann.rnames}{Additional rownames of \code{tab} to annotate; must be in \code{rownames(tab)}.}

\item{up.color}{Color for points that are upregulated (\code{logFC>0}).}

\item{down.color}{Color for points that are downregulated (\code{logFC<0}).}

\item{x.bound}{x-axis limits are set to \code{c(-x.bound, x.bound)}. If \code{NULL, x.bound=max(abs(tab[,lfc.col]))}.}

\item{y.bound}{y-axis limits are set to \code{c(0, y.bound)}. If \code{NULL, y.bound=max(tab[,'nlg10sig'])}.}

\item{type.sig}{Type of significance y-axis should use, either "p" or "FDR".}

\item{cut.color}{Color of points that meet both \code{cut.lfc} and \code{cut.sig}. If \code{NULL}, cutoffs are ignored.}

\item{cut.lfc}{Points need to have \code{|logFC| >= cut.lfc} to have \code{cut.color}.}

\item{cut.sig}{Points need to have significance \code{tab[,sig.col] <= cut.sig} to have \code{cut.color}.}

\item{sep}{Separator string between contrast names and suffix such as \code{logFC}.}

\item{na.lab}{Character vector of labels in \code{lab.col} to treat as missing, in addition to \code{NA}.}
}
\value{
A ggplot object, invisibly.
}
\description{
Volcano plot in ggplot2 using output from \code{ezlimma} package.
}
\details{
If \code{ntop.sig>0} or \code{ntop.lfc>0}, then \code{lab.col} must be in \code{colnames(tab)}.
}
