% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_covar_pca.R
\name{multi_covar_pca}
\alias{multi_covar_pca}
\title{PCA plots of top two components colored by each covariate}
\usage{
multi_covar_pca(object, pheno.df, name = "covar_pca", grp.var = "grp",
  covars = setdiff(colnames(pheno.df), grp.var), alpha = 1,
  all.size = NULL, facet = NULL, rm.leg.title = FALSE,
  labels = FALSE, manual.color = NULL, manual.shape = NULL,
  plot = TRUE)
}
\arguments{
\item{object}{Matrix-like object with features (e.g. genes) as rows and samples as columns.}

\item{pheno.df}{Dataframe with rows as samples and columns as phenotypes.}

\item{name}{Name of file to create. Set to \code{NA} to plot to screen instead of to file.}

\item{grp.var}{Column of \code{pheno.df} to include as group variable in PCA plots.}

\item{covars}{Columns of \code{pheno.df} to include as covariates in PCA plots.}

\item{alpha}{Transparency, passed to \code{\link[ggplot2]{geom_point}}.}

\item{all.size}{Passed to \code{\link[ggplot2]{geom_point}} \code{size} parameter to give size for all points without
appearing in legend. \code{ggplot2} default is size=2.}

\item{facet}{A formula with columns in \code{pheno.df} to facet by.}

\item{rm.leg.title}{Logical indicating if legend title should be removed.}

\item{labels}{Logical, should sample labels be added next to points?}

\item{manual.color}{Vector passed to \code{\link[ggplot2:scale_manual]{scale_colour_manual}} for creating a
discrete color scale. Vector length should be equal to number of levels in mapped variable.}

\item{manual.shape}{Vector passed to \code{\link[ggplot2:scale_manual]{scale_shape_manual}} for creating a
discrete color scale. Vector length should be equal to number of levels in mapped variable.}

\item{plot}{Logical; should plot be generated?}
}
\value{
Invisibly, a list of the first two principal components appended to \code{pheno.df} for each covariate.
}
\description{
PCA plot of top two principal components using \code{ggplot2} whose shape is defined by group and color in each plot
by iterating over each covariate.
}
\details{
PCA is calculated with \code{\link[stats]{prcomp}}.
}
