% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_heat.R
\name{multi_heat}
\alias{multi_heat}
\title{Plot one heatmap per comparison}
\usage{
multi_heat(
  tab,
  object,
  pheno.df = NULL,
  labrows = rownames(object),
  labcols = colnames(object),
  main = "Log2 Expression",
  name = "heats",
  sc = "ctr",
  clip = NA,
  color.v = NULL,
  unique.rows = FALSE,
  only.labrows = FALSE,
  ntop = 50,
  stat.tab = NULL,
  cutoff = 0.05,
  reorder_rows = TRUE,
  reorder_cols = FALSE,
  fontsize_row = 10,
  fontsize_col = 10,
  na.lab = c("---", ""),
  plot = TRUE,
  width = NA,
  height = NA,
  verbose = FALSE
)
}
\arguments{
\item{tab}{Table of output from \code{ezlimma}.}

\item{object}{Matrix-like object with features (e.g. genes) as rows and samples as columns.}

\item{pheno.df}{Data frame with rows as samples and columns as phenotypes.}

\item{labrows}{Labels for rows, e.g. gene symbols. This can be of length 1 (in which case it is recycled),
of length \code{nrow(object)}.}

\item{labcols}{Labels for columns. This can be of length 1 (in which case it is recycled),
of length \code{ncol(object)}.}

\item{main}{Main title of plot.}

\item{name}{Name of file to create. Set to \code{NA} to plot to screen instead of to file.}

\item{sc}{Row scaling. Should rows be centered ('ctr'), z-scored ('z'), or neither ('none').}

\item{clip}{Values with magnitude > \code{clip} are reset to value \code{clip}. If given, must be > 0.}

\item{color.v}{Color palette for heatmap. If \code{NULL}, it's set to
\code{colorRampPalette(rev(brewer.pal(n=9, name='RdYlBu')))(50)}.}

\item{unique.rows}{Logical; remove duplicated row labels to make rows unique?}

\item{only.labrows}{Logical; only include rows where \code{labrows} aren't missing
(missing is defined by \code{na.lab})?}

\item{ntop}{Scalar number of rows to include.}

\item{stat.tab}{Matrix-like object with statistics, such as p-values, per column, used to add "*". If given, its
dimensions should match \code{object} and it cannot have \code{NA}s.}

\item{cutoff}{Cutoff such that elements with \code{stats.tab < cutoff} show asterisk if \code{stats.tab} and
\code{cutoff} are not \code{NULL}.}

\item{reorder_rows}{Logical; should rows be reordered with hierarchical clustering?}

\item{reorder_cols}{Logical; should columns be reordered with hierarchical clustering?}

\item{fontsize_row}{Font size for row labels.}

\item{fontsize_col}{Font size for column labels.}

\item{na.lab}{Character vector of labels in \code{lab.col} to treat as missing, in addition to \code{NA}.}

\item{plot}{Logical; should plot be generated?}

\item{width}{Manual option for determining the output file width in inches.}

\item{height}{Manual option for determining the output file height in inches.}

\item{verbose}{Logical; print pruning messages to console?}
}
\description{
Plot one heatmap per comparison with \code{ezheat}, where features in \code{object} are reordered per comparison
using \code{tab}.
}
\details{
\code{rownames(tab)} and \code{rownames(object)} should overlap, \code{labrows} should correspond to \code{object}
and some \code{colnames(tab)} should end in \code{.p}, so they can be identified.

To prevent this function from being called with an unnamed \code{labrows} that corresponds to \code{tab} instead of \code{object},
which is incorrect, if \code{labrows} is not \code{names(object)} (the default) then it must be named.
}
