% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_pca.R
\name{multi_pca}
\alias{multi_pca}
\title{PCA plot of top two components}
\usage{
multi_pca(mat, pheno, name = "pca", scale. = FALSE, alpha = 1,
  all.size = NULL, facet = NULL, rm.leg.title = FALSE,
  add.labels = FALSE, manual.color = NULL, manual.shape = NULL, ...)
}
\arguments{
\item{mat}{Matrix with samples as columns}

\item{pheno}{Dataframe of phenotypes}

\item{name}{Name of PNG to write out. Set to \code{NA} to suppress writing to file.}

\item{scale.}{Logical indicating whether to scale in \code{\link[stats]{prcomp}}.}

\item{alpha}{Transparency, passed to \code{\link[ggplot2]{geom_point}}.}

\item{all.size}{Passed to \code{\link[ggplot2]{geom_point}} \code{size} parameter to give size for all points without appearing in legend. \code{ggplot2} default is size=2.}

\item{facet}{Variable to facet by.}

\item{rm.leg.title}{Logical indicating if legend title should be removed.}

\item{add.labels}{Logical indicating if sample labels should be added next to points.}

\item{manual.color}{Values passed to \code{\link[ggplot2]{scale_colour_manual}}.}

\item{manual.shape}{Values passed to \code{\link[ggplot2]{scale_shape_manual}}.}

\item{...}{Passed to \code{\link[ggplot2]{ggplot}} \code{aes_string} parameter.}
}
\description{
PCA plot of top two components using \code{ggplot2}.
}
