% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_venn.R
\name{multi_venn}
\alias{multi_venn}
\title{Multiple Venn diagrams}
\usage{
multi_venn(
  tab,
  prefix.lst,
  p.cutoff = NULL,
  fdr.cutoff = NULL,
  logfc.cutoff = NULL,
  circle.names.lst = prefix.lst,
  main.v = "",
  name = "multi",
  cex = c(1, 1, 1)
)
}
\arguments{
\item{tab}{Table of output from \code{ezlimma}.}

\item{prefix.lst}{A list of \code{prefix.v} vectors passed to \code{\link{ezvenn}}.}

\item{p.cutoff}{p-value cutoff to threshold features in \code{tab}. p-values < \code{p.cutoff} are significant.}

\item{fdr.cutoff}{FDR cutoff to threshold features in \code{tab}. FDRs < \code{fdr.cutoff} are significant.}

\item{logfc.cutoff}{log fold-change cutoff to threshold features in \code{tab}, when \code{tab} came from
\code{\link[ezlimma]{limma_contrasts}}.}

\item{circle.names.lst}{List of length \code{length(prefix.lst)}, where each element is a \code{circle.names} vector.
Passed to \code{\link{ezvenn}}.}

\item{main.v}{Vector of names of length \code{length(prefix.lst)}.}

\item{name}{Name of file to create. Set to \code{NA} to plot to screen instead of to file.}

\item{cex}{A numerical value giving the amount by which plotting text and symbols should be magnified relative to
the default. See \code{\link[graphics]{par}}.}
}
\value{
Invisibly, a matrix with elements {-1, 0, 1} indicating which features (rows) of \code{tab} were significant
with specified cutoffs. 0 indicates no significant change; -1 indicates down; and 1 indicates up if corresponding
\code{logFC} columns are found, otherwise 1 indicates significance.
}
\description{
Plot multiple Venn diagrams using output from \code{ezlimma} package. Each is a page in a PDF, and a single table
with all results is invisibly returned.
}
\details{
One of \code{fdr.cutoff} or \code{p.cutoff} must be given. If both are given, only \code{fdr.cutoff} is used.
\code{logfc.cutoff} if given is used in addition to these.

\code{"_venn.pdf"} is appended to \code{name}.

If \code{main.v==""} and \code{names(prefix.lst)} is not \code{NULL}, \code{main.v} is assigned
\code{names(prefix.lst)}.
}
