% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_volcano.R
\name{multi_volcano}
\alias{multi_volcano}
\title{Multiple Volcano plots}
\usage{
multi_volcano(tab, lab.col = NULL, ntop.sig = 0, ntop.lfc = 0,
  alpha = 0.4, name = "volcanoes", ann.rnames = NULL,
  up.ann.color = "black", down.ann.color = "black",
  same.scale = FALSE, type.sig = c("p", "FDR"), cut.color = NULL,
  cut.lfc = 1, cut.sig = 0.05, p05.line = FALSE, sep = ".",
  na.lab = c("---", ""), plot = TRUE)
}
\arguments{
\item{tab}{Table of output from \code{ezlimma}.}

\item{lab.col}{Column name or index of tab with labels, such as gene symbol, annotating features. If \code{NULL},
no points are annotated.}

\item{ntop.sig}{Number of top significant features to annotate.}

\item{ntop.lfc}{Number of top logFC features to annotate.}

\item{alpha}{Transparency for non-annotated points, passed to \code{\link[ggplot2]{geom_point}}.}

\item{name}{Name of file to create. Set to \code{NA} to plot to screen instead of to file.}

\item{ann.rnames}{Character vector of additional rownames of \code{tab} to annotate; elements must be in \code{rownames(tab)}.}

\item{up.ann.color}{Color for annotated points that are upregulated (\code{logFC>0}).}

\item{down.ann.color}{Color for annotated points that are downregulated (\code{logFC<0}).}

\item{same.scale}{Logical, should axes' scales be the same for the different plots?}

\item{type.sig}{Type of significance y-axis should use, either "p" or "FDR".}

\item{cut.color}{Color of points that meet both \code{cut.lfc} and \code{cut.sig}. If \code{NULL}, cutoffs are ignored.}

\item{cut.lfc}{Points need to have \code{|logFC| >= cut.lfc} to have \code{cut.color}.}

\item{cut.sig}{Points need to have significance \code{tab[,sig.col] <= cut.sig} to have \code{cut.color}.}

\item{p05.line}{Logical, should a dashed line at p=0.05 be added? If \code{TRUE}, \code{type.sig} should be \code{"p"}.}

\item{sep}{Separator string between contrast names and suffix such as \code{logFC}.}

\item{na.lab}{Character vector of labels in \code{lab.col} to treat as missing, in addition to \code{NA}.}

\item{plot}{Logical; should plot be generated?}
}
\value{
Invisibly, a list of ggplot objects from \code{\link{ezvolcano}}.
}
\description{
Multiple Volcano plots from a table output from \code{ezlimma}.
}
\details{
Uses \code{colnames(tab)} that have suffix \code{logFC} to infer comparisons.
}
