% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighbor_nms.R
\name{neighbor_nms}
\alias{neighbor_nms}
\title{Get names of nodes' neighbors}
\usage{
neighbor_nms(gr, nodes)
}
\arguments{
\item{gr}{graph object of class \code{igraph}.}

\item{nodes}{Node names in the graph}
}
\description{
Get node names of nodes' direct neighbors
}
