% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/order_stats_by_alternative.R
\name{order_stats_by_alternative}
\alias{order_stats_by_alternative}
\title{Order values based on alternative hypothesis}
\usage{
order_stats_by_alternative(stat.v, alternative)
}
\arguments{
\item{stat.v}{Numeric vector of statistics to order; preferably a named vector.}

\item{alternative}{Alternative hypothesis; must be one of \code{"two.sided"}, \code{"greater"} or \code{"less"}. 
\code{"greater"} corresponds to positive association, \code{"less"} to negative association.}
}
\description{
Order values based on alternative hypothesis.
}
