% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pivot_roast_longer.R
\name{pivot_roast_longer}
\alias{pivot_roast_longer}
\title{Lengthen output table from \code{ezlimma::roast_*} to be plotted with \code{ggplot2}}
\usage{
pivot_roast_longer(tab, prefix.v, direction.v = c("Up", "Down", "Mixed"))
}
\arguments{
\item{tab}{Table of output from \code{ezlimma::roast_*}.}

\item{prefix.v}{Character vector of prefixes.}

\item{direction.v}{Character vector of directions to include in output.}
}
\value{
Table of pathway statistics to be plotted using \code{ggplot2}.
}
\description{
Lengthen output table from \code{ezlimma::roast_*} to be plotted with \code{ggplot2} in \code{ezlimmaplot}.
Name is based on \code{tidyr::pivot_longer} function.
}
