% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_man.R
\name{plot_man}
\alias{plot_man}
\title{Plot mediation exposure, mediator, and outcome per mediator}
\usage{
plot_man(
  E,
  M,
  Y,
  name = "top_mediators",
  main.v = "",
  xlab = "Log2 abundance",
  ylab = "Outcome",
  check.names = TRUE,
  manual.color = NULL
)
}
\arguments{
\item{E}{A numeric vector of exposures.}

\item{M}{A numeric matrix-like data object with one row per feature and one column per sample of mediators.}

\item{Y}{A numeric vector of \code{length(E)} of outcomes. Only continuous, normally distributed outcomes
currently supported.}

\item{name}{Name of file to create. Set to \code{NA} to plot to screen instead of to file.}

\item{main.v}{Character vector of main titles for plot.}

\item{xlab}{Label for x-axis.}

\item{ylab}{Label for y-axis.}

\item{check.names}{Logical; should \code{names(E)==colnames(M) & colnames(M)==names(Y)} be checked?}

\item{manual.color}{Vector passed to both \code{\link[ggplot2:scale_manual]{scale_fill_manual}} and
\code{\link[ggplot2:scale_manual]{scale_color_manual}}.}
}
\value{
Invisibly, a \pkg{ggplot2} object from the last row of \code{M} that was plotted.
}
\description{
Plot mediation exposure, mediator, and outcome per mediator in \code{M}, as used in \code{\link[Hitman]{lotman}} &
\code{\link[Hitman]{hitman}}, using \pkg{ggplot2}.
}
\details{
If \code{manual.color} is given, its length should be \code{length(unique(E))}.
}
