% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pwy.R
\name{plot_pwy}
\alias{plot_pwy}
\title{Plot network diagram for a pathway}
\usage{
plot_pwy(feat.tab, G.pwy, gr, stat.colnm, annot.col = NULL, ntop = 7,
  name = NULL, colorbar.nm = stat.colnm, alternative = c("two.sided",
  "greater", "less"), repel = FALSE, plot = TRUE, seed = 0)
}
\arguments{
\item{feat.tab}{Matrix-like object for all analytes with rownames as analyte names, which should
have some overlap with \code{G.pwy$genes}. 1st column with stats.
Stats should be finite unless they are \code{NA}. Stats for all nodes are used to define color scale,
for consistency across pathway visualizations. 1st column name used as stat descriptor in plot.}

\item{G.pwy}{Element of list object \code{G} used in \pkg{ezlimma} or \pkg{{Pame}}. Should have elements
\code{name}, \code{description}, \code{genes}. Accessible via \code{G[[pathway.name]]}.}

\item{gr}{graph object of class \code{igraph}.}

\item{stat.colnm}{Column name in \code{feat.tab} indicating scores to select top analytes for this pathway and plot.}

\item{annot.col}{Column index or name in \code{feat.tab} of annotations to replace \code{rownames(G.pwy)}.
\code{NA}s in \code{annot.col} are ignored, and these nodes retain their rowname as a label.}

\item{ntop}{Number of top impactful analytes to plot, >=2. Their network neighbors may also be included.}

\item{name}{Name of file to plot to. If \code{NULL}, a filename is created using \code{colnames(Gmat.pwy)[1]}.
Set to \code{NA} to plot to screen instead of to file.}

\item{colorbar.nm}{Title of color bar.}

\item{alternative}{Alternative hypothesis; must be one of \code{"two.sided"}, \code{"greater"} or \code{"less"}. 
\code{"greater"} corresponds to positive association, \code{"less"} to negative association.}

\item{repel}{Logcal; use the repel functionality from \pkg{ggrepel} to avoid overlapping text?}

\item{plot}{Logical; should plot be generated?}

\item{seed}{Integer seed to set for reproducility if \code{fun="mroast"}, since \code{mroast} uses random 
simulations. Ignored if \code{fun="fry"}.}
}
\value{
Invisibly, a \code{\link[tidygraph]{tbl_graph}}, a subclass of
\pkg{igraph} so every \pkg{igraph} method will work as expected.
}
\description{
Plot nodes most impacting a pathway as a network diagram with node color corresponding to z-score.
The most impactful nodes are inferred by assuming the input here is the same as
was used to calculate pathway significance.
}
\details{
Unmeasured nodes have stat of \code{NA} and are drawn gray.
}
