% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pwy.R
\name{plot_pwy_network}
\alias{plot_pwy_network}
\title{Plot network diagram for a pathway}
\usage{
plot_pwy_network(
  feat.tab,
  G.pwy,
  gr,
  stat.colnm,
  annot.colnm,
  ntop = 7,
  name = NULL,
  colorbar.nm = stat.colnm,
  alternative = c("two.sided", "greater", "less"),
  repel = FALSE,
  plot = TRUE,
  seed = 0
)
}
\arguments{
\item{feat.tab}{Matrix-like object for all analytes with row names as analyte names, which should
have some overlap with \code{G.pwy$genes}.}

\item{G.pwy}{Element of list object \code{G} used in \pkg{ezlimma}. Should have elements
\code{name}, \code{description}, \code{genes}. Accessible via \code{G[[pathway.name]]}.}

\item{gr}{graph object of class \code{igraph}, possibly a result of \code{\link{edgelist2graph}}.}

\item{stat.colnm}{Column name in \code{feat.tab} indicating scores to select top analytes for this pathway and plot.
Stats should be finite unless they are \code{NA}. Stats for all nodes are used to define color scale,
for consistency across pathway visualizations.}

\item{annot.colnm}{Column name in \code{feat.tab} of node names to possibly replace \code{rownames(feat.tab)}.
Set this column to be \code{rownames(feat.tab)} if these are desired node names. Must not be \code{NA}.}

\item{ntop}{Number of top impactful analytes to plot, >=2. Their network neighbors may also be included.}

\item{name}{Name of file to plot to. If \code{NULL}, a filename is created using \code{colnames(Gmat.pwy)[1]}.
Set to \code{NA} to plot to screen instead of to file.}

\item{colorbar.nm}{Title of color bar.}

\item{alternative}{Alternative hypothesis; must be one of \code{"two.sided"}, \code{"greater"} or \code{"less"}. 
\code{"greater"} corresponds to positive association, \code{"less"} to negative association.}

\item{repel}{Logcal; use the repel functionality from \pkg{ggrepel} to avoid overlapping text?}

\item{plot}{Logical; should plot be generated?}

\item{seed}{Numeric seed value for reproducibility.}
}
\value{
Invisibly, a list with \code{\link[tidygraph]{tbl_graph}}, a subclass of
\pkg{igraph} so every \pkg{igraph} method will work as expected. If \code{plot=TRUE}, there
is a second element of the list, which is a \code{ggplot} object.
}
\description{
Plot nodes most impacting a pathway as a network diagram with node color corresponding to z-score.
The most impactful nodes are inferred by assuming the input here is the same as
was used to calculate pathway significance. The \code{feat.tab} is assumed to have all nodes to consider
for plotting. Including nodes with \code{NA} stats allows these to be included in networks as a connector.
An \code{annot.colnm} should have preferred labels for all nodes.
}
\details{
Unmeasured nodes have stat of \code{NA} and are drawn gray.
}
