% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sif2edgelist.R
\name{sif2edgelist}
\alias{sif2edgelist}
\title{Clean Pathway Commons SIF file}
\usage{
sif2edgelist(sif, rm.ids = "CHEBI:15377")
}
\arguments{
\item{sif}{SIF file from Pathway Commons}

\item{rm.ids}{vector of analyte IDs to remove. Default \code{CHEBI:15377} corresponds to water.}
}
\value{
Edge-list matrix with two columns.
}
\description{
Remove some ubiquitious IDs from Pathway Commons SIF file, and transform to edge-list.
}
