% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signif_hist.R
\name{signif_hist}
\alias{signif_hist}
\title{Plot histograms of significance (p-value & FDR) columns}
\usage{
signif_hist(tab, p.suffix = "p", fdr.suffix = "FDR", sep = ".",
  pi0 = FALSE, name = "signif_hist", plot = TRUE)
}
\arguments{
\item{tab}{Table of output from \code{ezlimma}.}

\item{p.suffix}{Suffix for p-value columns. P-value column names cannot be duplicated.}

\item{fdr.suffix}{Suffix for FDR columns. Set to \code{NA} if no FDR columns. FDR column names cannot be duplicated.}

\item{sep}{Separator for column names before \code{p} or \code{FDR}.}

\item{pi0}{Logical indicating if proportion of null hypotheses should be calculated per p-value histogram. If
\code{TRUE}, \code{limma::propTrueNull} is used, so \code{limma} package is needed.}

\item{name}{Name of PDF file to write. Set to \code{NA} to suppress writing to PDF.}

\item{plot}{Logical; if \code{FALSE} no plot is generated.}
}
\value{
Invisibly, a subset of \code{tab} with only columns that contain significances.
}
\description{
Plot histograms of significance (p-value & FDR) columns.
}
\details{
Some p-value columns must be identifiable using \code{p.suffix}. If \code{!is.na(fdr.suffix)}, FDR
colnames must have same prefix.
}
