% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sub_labels.R
\name{sub_labels}
\alias{sub_labels}
\title{Replace names with non-missing label.v}
\usage{
sub_labels(nms, label.v, na.lab = c("---", ""))
}
\arguments{
\item{nms}{Vector of names, e.g. rownames.}

\item{label.v}{Vector of label.v, e.g. \code{labrows}.}

\item{na.lab}{Character vector of labels in \code{lab.col} to treat as missing, in addition to \code{NA}.}
}
\value{
Vector of replaced rownames; vector's names are rownames
}
\description{
Replace names, esp. rownames, with non-missing label.v. \code{nms} and \code{label.v} must be of same length,
and \code{nms} must not have \code{NA}s.
}
