% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_prefix.R
\name{extract_prefix}
\alias{extract_prefix}
\title{Extract prefix from names}
\usage{
extract_prefix(nm, suffix = "p", sep = ".")
}
\arguments{
\item{nm}{Character vector of names to search.}

\item{suffix}{Character string to search for at end of \code{nm}.}

\item{sep}{Separator in \code{nm} preceeding \code{suffix}. One of \code{".", "_", NA}, If \code{NA}, elements
of \code{nm} must match \code{suffix} exactly.}
}
\value{
A character vector of prefixes. If there is no prefix and no sep, \code{NA} is returned.
}
\description{
Extract prefixes from names, presumably colnames. For example, the prefix of \code{"x.p"} would be \code{"x"}.
}
