% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ezvenn.R
\name{ezvenn}
\alias{ezvenn}
\title{Venn diagrams}
\usage{
ezvenn(
  tab,
  prefix.v = NULL,
  p.cutoff = NULL,
  fdr.cutoff = NULL,
  logfc.cutoff = NULL,
  include = c("up", "down"),
  circle.names = prefix.v,
  main = "",
  name = NA,
  cex = c(1, 1, 1),
  plot = TRUE
)
}
\arguments{
\item{tab}{Table of output from \code{ezlimma}.}

\item{prefix.v}{A vector of prefixes that prefix \code{.p}, \code{.FDR}, or \code{.logFC} in \code{colnames(tab)}.
These will be the columns that are used in the Venn diagram. If \code{NULL} these are inferred from
\code{colnames(tab)} that end with \code{.p}.}

\item{p.cutoff}{p-value cutoff to threshold features in \code{tab}. p-values < \code{p.cutoff} are significant.}

\item{fdr.cutoff}{FDR cutoff to threshold features in \code{tab}. FDRs < \code{fdr.cutoff} are significant.}

\item{logfc.cutoff}{log fold-change cutoff to threshold features in \code{tab}, when \code{tab} came from
\code{\link[ezlimma]{limma_contrasts}}.}

\item{include}{Character vector. If \code{logFC} columns are found, this specifies whether all differentially
expressed genes should be counted, or whether the counts should be restricted to genes changing in a certain direction.
Choices are "both" for all differentially expressed genes, "up" for up-regulated genes only or
"down" for down-regulated genes only. If include=c("up","down") then both the up and down counts will be shown.
If \code{logFC} columns are not found, \code{include = "both"}.}

\item{circle.names}{Vector of circle names corresponding to \code{prefix.v}.}

\item{main}{Main title of plot.}

\item{name}{Name of file to create. Set to \code{NA} to plot to screen instead of to file.}

\item{cex}{A numerical value giving the amount by which plotting text and symbols should be magnified relative to
the default. See \code{\link[graphics]{par}}.}

\item{plot}{Logical; should plot be generated?}
}
\value{
Invisibly, a matrix with elements {-1, 0, 1} indicating which features (rows) of \code{tab} were significant
with specified cutoffs. 0 indicates no significant change; -1 indicates down; and 1 indicates up if corresponding
\code{logFC} columns are found, otherwise 1 indicates significance.
}
\description{
Venn diagrams using output from \code{ezlimma} package.
}
\details{
One of \code{fdr.cutoff} or \code{p.cutoff} must be given. If both are given, only \code{fdr.cutoff} is used.
\code{logfc.cutoff} if given is used in addition to these.

\code{"_venn.pdf"} is appended to \code{name}.
}
