% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_volcano.R
\name{multi_volcano}
\alias{multi_volcano}
\title{Multiple Volcano plots}
\usage{
multi_volcano(tab, lab.col = NULL, ntop.sig = 0, ntop.lfc = 0,
  name = "volcanoes", ann.rnames = NULL, up.ann.color = "black",
  down.ann.color = "black", same.scale = FALSE, type.sig = c("p",
  "FDR"), cut.color = NULL, cut.lfc = 1, cut.sig = 0.05, sep = ".",
  na.lab = c("---", ""))
}
\arguments{
\item{tab}{Table of output from \code{ezlimma}.}

\item{lab.col}{Column name or index of tab with labels, such as gene symbol, annotating features. If \code{NULL},
no points are annotated.}

\item{ntop.sig}{Number of top significant features to annotate.}

\item{ntop.lfc}{Number of top logFC features to annotate.}

\item{name}{Name of file to create. Set to \code{NA} to plot to screen instead of to PDF.}

\item{ann.rnames}{Additional rownames of \code{tab} to annotate; must be in \code{rownames(tab)}.}

\item{up.ann.color}{Color for annotated points that are upregulated (\code{logFC>0}).}

\item{down.ann.color}{Color for annotated points that are downregulated (\code{logFC<0}).}

\item{type.sig}{Type of significance y-axis should use, either "p" or "FDR".}

\item{cut.color}{Color of points that meet both \code{cut.lfc} and \code{cut.sig}. If \code{NULL}, cutoffs are ignored.}

\item{cut.lfc}{Points need to have \code{|logFC| >= cut.lfc} to have \code{cut.color}.}

\item{cut.sig}{Points need to have significance \code{tab[,sig.col] <= cut.sig} to have \code{cut.color}.}

\item{sep}{Separator string between contrast names and suffix such as \code{logFC}.}

\item{na.lab}{Character vector of labels in \code{lab.col} to treat as missing, in addition to \code{NA}.}
}
\value{
Invisibly, a list of ggplot objects from \code{\link{ezvolcano}}.
}
\description{
Multiple Volcano plots from a table output from \code{ezlimma}.
}
\details{
Uses \code{colnames(tab)} that have suffix \code{logFC} to infer comparisons.
}
