% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prune_mat.R
\name{prune_mat}
\alias{prune_mat}
\title{Prune matrix}
\usage{
prune_mat(object, symbols = NULL, only.symbols = FALSE,
  unique.rows = FALSE, ntop = NULL, na.lab = c("---", ""),
  verbose = TRUE)
}
\arguments{
\item{object}{Matrix-like object.}

\item{symbols}{Labels corresponding to rows, e.g. gene symbols.}

\item{only.symbols}{Logical, if \code{TRUE} only include rows with a non-missing symbol.}

\item{unique.rows}{Logical, indicating if to remove duplicated row labels to make rows unique.}

\item{ntop}{Scalar number of rows to include.}

\item{na.lab}{Character vector of labels in \code{lab.col} to treat as missing, in addition to \code{NA}.}

\item{verbose}{Logical indicating whether to print messages to console.}
}
\value{
Pruned matrix.
}
\description{
Prune (subset) matrix before plotting as heatmap.
}
