% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signif_hist.R
\name{signif_hist}
\alias{signif_hist}
\title{Plot histograms of significance (p-value & FDR) columns}
\usage{
signif_hist(
  tab,
  p.suffix = "p",
  fdr.suffix = "FDR",
  sep = ".",
  pi0 = FALSE,
  name = "signif_hist",
  plot = TRUE
)
}
\arguments{
\item{tab}{Table of output from \code{ezlimma}.}

\item{p.suffix}{Suffix for p-value columns. P-value column names cannot be duplicated.}

\item{fdr.suffix}{Suffix for FDR columns. Set to \code{NA} if no FDR columns. FDR column names cannot be duplicated.}

\item{sep}{Separator for column names before \code{p} or \code{FDR}, passed to \code{\link{extract_prefix}}.
If not found, it is assumed to be \code{NA}.}

\item{pi0}{Logical indicating if proportion of null hypotheses should be calculated per p-value histogram. If
\code{TRUE}, \code{\link[limma]{propTrueNull}} with \code{method="convest"} is used.}

\item{name}{Name of file to create. Set to \code{NA} to plot to screen instead of to file.}

\item{plot}{Logical; should plot be generated?}
}
\value{
Invisibly, a subset of \code{tab} with only columns that contain significances.
}
\description{
Plot histograms of significance (p-value & FDR) columns.
}
\details{
Some p-value columns must be identifiable using \code{p.suffix}. If \code{!is.na(fdr.suffix)}, FDR
colnames must have same prefix.
}
