% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sub_labels.R
\name{sub_labels}
\alias{sub_labels}
\title{Replace names with non-missing labels}
\usage{
sub_labels(nms, labels, na.lab = c("---", ""))
}
\arguments{
\item{nms}{Vector of names, e.g. rownames.}

\item{labels}{Vector of labels, e.g. \code{labrows}.}

\item{na.lab}{Character vector of labels in \code{lab.col} to treat as missing, in addition to \code{NA}.}
}
\value{
Vector of replaced rownames; vector's names are rownames
}
\description{
Replace names, esp. rownames, with non-missing labels.
}
